+++ Assets/Scripts/BeadsProviders/ILinkBeadsProvider.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace BeadsProviders
+{
+    public interface ILinkBeadsProvider
+    {
+        List<Vector3[]> GetBeadsList();
+    }
+}
