+++ Assets/Scripts/Domain/Bead.cs.meta
