+++ Assets/Scripts/Domain/Component.cs
@@ -1,0 +1,14 @@
+using System.Collections.Generic;
+
+namespace Domain
+{
+    public class Component
+    {
+        public readonly List<Bead> BeadList;
+
+        public Component(List<Bead> beadList)
+        {
+            BeadList = beadList;
+        }
+    }
+}
